Fix payment page
Fix payment page
Re-arrange layout HTML

+++ HuloToys-Front-End/wwwroot/modules/global.js
@@ -232,27 +232,32 @@ var global_service = {
                         && !img_src.includes("http"))
                         img_src = API_URL.StaticDomain + item.avatar
                     var amount_html = 'Giá liên hệ'
+                    var has_price=false
                     if (item.amount_max != undefined
                         && item.amount_max != null
                         && item.amount_min != undefined
                         && item.amount_min != null) {
                         amount_html = global_service.Comma(item.amount_min) + ' - ' + global_service.Comma(item.amount_max)
+                        has_price=true
                     }
                     else if (item.amount != undefined
                         && item.amount != null && item.amount > 0) {
                         amount_html = global_service.Comma(item.amount)
+                        has_price=true
+                    }
+                    if (has_price) {
+                        html += HTML_CONSTANTS.Home.SlideProductItem
+                            .replaceAll('{url}', '/san-pham/' + global_service.RemoveUnicode(item.name).replaceAll(' ', '-') + '--' + item._id)
+                            .replaceAll('{avt}', img_src)
+                            .replaceAll('{name}', item.name)
+                            .replaceAll('{amount}', amount_html)
+                            //.replaceAll('{review_point}', (item.rating == null || item.rating == undefined || item.rating <= 0) ? '5' : item.rating)
+                            .replaceAll('{review_point}', '5')
+                            .replaceAll('{review_count}', '')
+                            .replaceAll('{old_price_style}', '')
+                            .replaceAll('{price}', (item.amount == null || item.amount == undefined || item.amount <= 0) ? global_service.Comma(item.amount) + ' đ' : '')
 
                     }
-                    html += HTML_CONSTANTS.Home.SlideProductItem
-                        .replaceAll('{url}', '/san-pham/' + global_service.RemoveUnicode(item.name).replaceAll(' ', '-') + '--' + item._id)
-                        .replaceAll('{avt}', img_src)
-                        .replaceAll('{name}', item.name)
-                        .replaceAll('{amount}', amount_html)
-                        //.replaceAll('{review_point}', (item.rating == null || item.rating == undefined || item.rating <= 0) ? '5' : item.rating)
-                        .replaceAll('{review_point}', '5')
-                        .replaceAll('{review_count}', '')
-                        .replaceAll('{old_price_style}', '')
-                        .replaceAll('{price}', (item.amount == null || item.amount == undefined || item.amount <= 0) ? global_service.Comma(item.amount) + ' đ' : '')
                 });
             }
             element.html(html)
